Template System Improvements
This no longer creates a file on the server for exporting, rather it downloads it directly.  This makes one tab redundant so it is removed.  If a user goes to try and mass edit templates on a theme and it doesn't have any of the socialgroups templates, it creates them based off what is in inc/plugins/socialgroups/templates.json.  Makes some form headers look better as well.